refine: address pr comments
