+++ package.json
@@ -72,8 +72,6 @@
     "aws-jwt-verify": "^5.1.0",
     "express": "^5.1.0",
     "fs": "^0.0.1-security",
-    "jose": "^6.0.11",
-    "jsonwebtoken": "^9.0.2",
-    "node-cron": "^4.2.1"
+    "jose": "^6.0.11"
   }
 }
